break up functions and files
